Update anat/inu_correction.sh, anat/brain_extraction.sh, anat/map_jacobian.sh, anat/map_myelin.sh, anat/rician_denoise.sh, anat/rigid_alignment.sh, anat/template_normalization.sh, anat/tissue_segmentation.sh files

+++ anat/inu_correction.sh
@@ -8,31 +8,54 @@
 # Date: 2020-02-26
 # Software: FSL
 #===============================================================================
+PROC_START=$(date +%Y-%m-%dT%H:%M:%S%z)
+FCN_NAME=(`basename "$0"`)
+DATE_SUFFIX=$(date +%Y%m%dT%H%M%S%N)
+OPERATOR=$(whoami)
+DEBUG=false
+NO_LOG=false
+
+# actions on exit, write to logs, clean scratch
+function egress {
+  EXIT_CODE=$?
+  LOG_STRING=`date +"${OPERATOR}\t${FCN_NAME}\t${PROC_START}\t%Y-%m-%dT%H:%M:%S%z\t${EXIT_CODE}"`
+  if [[ "${NO_LOG}" == "false" ]]; then
+    FCN_LOG=/Shared/inc_scratch/log/benchmark_${FCN_NAME}.log
+    if [[ ! -f ${FCN_LOG} ]]; then
+      echo -e 'operator\tfunction\tstart\tend\texit_status' > ${FCN_LOG}
+    fi
+    echo -e ${LOG_STRING} >> ${FCN_LOG}
+    if [[ -v DIR_PROJECT ]]; then
+      PROJECT_LOG=${DIR_PROJECT}/log/${PREFIX}.log
+      if [[ ! -f ${PROJECT_LOG} ]]; then
+        echo -e 'operator\tfunction\tstart\tend\texit_status' > ${PROJECT_LOG}
+      fi
+      echo -e ${LOG_STRING} >> ${PROJECT_LOG}
+    fi
+  fi
+  if [[ "${DEBUG}" == "false" ]]; then
+    if [[ -d ${DIR_SCRATCH} ]]; then
+      if [[ "$(ls -A ${DIR_SCRATCH})" ]]; then
+        rm -R ${DIR_SCRATCH}/*
+      fi
+      rmdir ${DIR_SCRATCH}
+    fi
+  fi
+}
+trap egress EXIT
 
 # Parse inputs -----------------------------------------------------------------
-OPTS=`getopt -o hvdkl --long group:,prefix:,\
+OPTS=`getopt -o hdvkl --long group:,prefix:,\
 dimension:,image:,method:,mask:,smooth-kernel:,weight:,shrink:,convergence,bspline:,hist-sharpen:,\
 dir-save:,dir-scratch:,dir-code:,dir-pincsource:,\
-help,verbose,keep,no-log -n 'parse-options' -- "$@"`
+help,debug,verbose,keep,no-log -n 'parse-options' -- "$@"`
 if [ $? != 0 ]; then
   echo "Failed parsing options" >&2
   exit 1
 fi
 eval set -- "$OPTS"
 
-# actions on exit, e.g., cleaning scratch on error ----------------------------
-function egress {
-  if [[ -d ${DIR_SCRATCH} ]]; then
-    if [[ "$(ls -A ${DIR_SCRATCH})" ]]; then
-      rm -R ${DIR_SCRATCH}/*
-    fi
-    rmdir ${DIR_SCRATCH}
-  fi
-}
-trap egress EXIT
-
 # Set default values for function ---------------------------------------------
-DATE_SUFFIX=$(date +%Y%m%dT%H%M%S%N)
 GROUP=
 PREFIX=
 DIM=3
@@ -46,24 +69,24 @@ CONVERGENCE=[50x50x50x50,0.0]
 BSPLINE=[200,3]
 HIST_SHARPEN=[0.15,0.01,200]
 DIR_SAVE=
-DIR_SCRATCH=/Shared/inc_scratch/scratch_${DATE_SUFFIX}
+DIR_SCRATCH=/Shared/inc_scratch/${OPERATOR}_${DATE_SUFFIX}
 DIR_CODE=/Shared/inc_scratch/code
 DIR_PINCSOURCE=/Shared/pinc/sharedopt/apps/sourcefiles
 HELP=false
 VERBOSE=0
 KEEP=false
-NO_LOG=false
 
 while true; do
   case "$1" in
     -h | --help) HELP=true ; shift ;;
+    -d | --debug) DEBUG=true ; shift ;;
     -v | --verbose) VERBOSE=1 ; shift ;;
-    -d | --dimension) DIM="$2" ; shift 2 ;;
     -k | --keep) KEEP=true ; shift ;;
     -l | --no-log) NO_LOG=true ; shift ;;
     --group) GROUP="$2" ; shift 2 ;;
     --prefix) PREFIX="$2" ; shift 2 ;;
-    --image) IMAGE+="$2" ; shift 2 ;;
+    --dimension) DIM="$2" ; shift 2 ;;
+    --image) IMAGE="$2" ; shift 2 ;;
     --method) METHOD="$2" ; shift 2 ;;
     --mask) MASK="$2" ; shift 2 ;;
     --smooth-kernel) SMOOTH_KERNEL="$2" ; shift 2 ;;
@@ -92,17 +115,18 @@ if [[ "${HELP}" == "true" ]]; then
   echo '------------------------------------------------------------------------'
   echo "Usage: ${FUNC_NAME}"
   echo '  -h | --help              display command help'
+  echo '  -d | --debug             keep scratch folder for debugging'
   echo '  -v | --verbose           add verbose output to log file'
   echo '  -l | --no-log            disable writing to output log'
   echo '  --group <value>          group permissions for project,'
   echo '                           e.g., Research-kosciklab'
   echo '  --prefix <value>         prefix for output,'
   echo '                           default: sub-123_ses-1234abcd'
-  echo '  -d | --dimension <value> image dimension, 3=3D (default) or 4=4D'
+  echo '  --dimension <value>      image dimension, 3=3D (default) or 4=4D'
   echo '                           T1T2 method only works on 3D images.'
-  echo '  --image <value>          full path to image, multiple images allowed'
-  echo '                           if using T1T2, image 1 must be T1w, image 2'
-  echo '                           must be T2w'
+  echo '  --image <value>          full path to image, if using T1T2, input must'
+  echo '                           be a comma separted string for T1w and T2w'
+  echo '                           images, image 1 must be T1w, image 2 must be T2w'
   echo '  --method <value>         one of N4 or T1T2 (case insensitive)'
   echo '  --mask <value>           full path to region mask'
   echo '  --smooth-kernel <value>  smoothing kernel size in mm, default: 5'
@@ -124,17 +148,16 @@ if [[ "${HELP}" == "true" ]]; then
   echo '  --dir-pincsource <value> directory for PINC sourcefiles'
   echo '                       default: /Shared/pinc/sharedopt/apps/sourcefiles'
   echo ''
+  NO_LOG=true
   exit 0
 fi
 
 # Set up BIDs compliant variables and workspace --------------------------------
-proc_start=$(date +%Y-%m-%dT%H:%M:%S%z)
-
 DIR_PROJECT=`${DIR_CODE}/bids/get_dir.sh -i ${IMAGE[0]}`
 SUBJECT=`${DIR_CODE}/bids/get_field.sh -i ${IMAGE[0]} -f "sub"`
 SESSION=`${DIR_CODE}/bids/get_field.sh -i ${IMAGE[0]} -f "ses"`
 if [ -z "${PREFIX}" ]; then
-  PREFIX=`${DIR_CODE}/bids/get_bidsbase -s -i ${IMAGE[0]}`
+  PREFIX=`${DIR_CODE}/bids/get_bidsbase.sh -s -i ${IMAGE[0]}`
 fi
 
 if [ -z "${DIR_SAVE}" ]; then
@@ -143,9 +166,11 @@ fi
 mkdir -p ${DIR_SCRATCH}
 mkdir -p ${DIR_SAVE}
 
-# =============================================================================
+#===============================================================================
 # Start of Function
-# =============================================================================
+#===============================================================================
+IMAGE=(${IMAGE//,/ })
+
 if [[ "${METHOD,,}" == "t1t2" ]]; then
   # Form sqrt(T1w*T2w), mask this and normalise by the mean
   fslmaths ${IMAGE[0]} -mul ${IMAGE[1]} -abs -sqrt \
@@ -205,46 +230,35 @@ if [[ "${METHOD,,}" == "t1t2" ]]; then
 fi
 
 if [[ "${METHOD,,}" == "n4" ]]; then
-  NUM_IMAGE=${#IMAGE[@]}
-  for (( i=0; i<${NUM_IMAGE}; i++ )); do
-    # gather modality for output
-    MOD=(${IMAGE[${i}]})
-    MOD=(`basename "${MOD%.nii.gz}"`)
-    MOD=(${MOD##*_})
-
-    n4_fcn="N4BiasFieldCorrection"
-    n4_fcn="${n4_fcn} -d ${DIM}"
-    n4_fcn="${n4_fcn} -i ${IMAGE[${i}]}"
-    if [ -n "${MASK}" ]; then
-      n4_fcn="${n4_fcn} -x ${MASK}"
-    fi
-    if [ -n "${WEIGHT}" ]; then
-      n4_fcn="${n4_fcn} -w ${WEIGHT}"
-    fi
-    n4_fcn="${n4_fcn} -r ${RESCALE}"
-    n4_fcn="${n4_fcn} -s ${SHRINK}"
-    n4_fcn="${n4_fcn} -c ${CONVERGENCE}"
-    n4_fcn="${n4_fcn} -b ${BSPLINE}"
-    n4_fcn="${n4_fcn} -t ${HIST_SHARPEN}"
-    n4_fcn="${n4_fcn} -o [${DIR_SCRATCH}/${PREFIX}_prep-bias+N4_${MOD}.nii.gz,${DIR_SCRATCH}/${PREFIX}_prep-bias+N4+field_${MOD}.nii.gz]"
-    n4_fcn="${n4_fcn} -v ${VERBOSE}"
-    eval ${n4_fcn}
-    
-    mv ${DIR_SCRATCH}/${PREFIX}_prep-bias+N4_${MOD}.nii.gz ${DIR_SAVE}/
-    if [[ "${KEEP}" == "true" ]]; then
-      mv ${DIR_SCRATCH}/${PREFIX}_prep-bias+N4+field_${MOD}.nii.gz ${DIR_SAVE}/
-    fi
-  done
+  # gather modality for output
+  MOD=`${DIR_CODE}/bids/get_field.sh -i ${IMAGE[0]} -f "modality"`
+  
+  n4_fcn="N4BiasFieldCorrection"
+  n4_fcn="${n4_fcn} -d ${DIM}"
+  n4_fcn="${n4_fcn} -i ${IMAGE[${i}]}"
+  if [ -n "${MASK}" ]; then
+    n4_fcn="${n4_fcn} -x ${MASK}"
+  fi
+  if [ -n "${WEIGHT}" ]; then
+    n4_fcn="${n4_fcn} -w ${WEIGHT}"
+  fi
+  n4_fcn="${n4_fcn} -r ${RESCALE}"
+  n4_fcn="${n4_fcn} -s ${SHRINK}"
+  n4_fcn="${n4_fcn} -c ${CONVERGENCE}"
+  n4_fcn="${n4_fcn} -b ${BSPLINE}"
+  n4_fcn="${n4_fcn} -t ${HIST_SHARPEN}"
+  n4_fcn="${n4_fcn} -o [${DIR_SCRATCH}/${PREFIX}_prep-bias+N4_${MOD}.nii.gz,${DIR_SCRATCH}/${PREFIX}_prep-bias+N4+field_${MOD}.nii.gz]"
+  n4_fcn="${n4_fcn} -v ${VERBOSE}"
+  eval ${n4_fcn}
+  
+  mv ${DIR_SCRATCH}/${PREFIX}_prep-bias+N4_${MOD}.nii.gz ${DIR_SAVE}/
+  if [[ "${KEEP}" == "true" ]]; then
+    mv ${DIR_SCRATCH}/${PREFIX}_prep-bias+N4+field_${MOD}.nii.gz ${DIR_SAVE}/
+  fi
 fi
 
 #===============================================================================
 # End of Function
 #===============================================================================
-# Write log entry on conclusion ------------------------------------------------
-if [[ "${NO_LOG}" == "false" ]]; then
-  LOG_FILE=${DIR_PROJECT}/log/${PREFIX}.log
-  date +"task:$0,start:"${proc_start}",end:%Y-%m-%dT%H:%M:%S%z" >> ${LOG_FILE}
-fi
-
 exit 0
